Add files via upload
This program checks if a user-entered 15 or 16-digit credit card number is valid using the Luhn algorithm (a checksum formula used to validate credit card numbers).  It also identifies the type of credit card (e.g., Visa, MasterCard, American Express)  based on the first few digits of the card number.